mdevctl: Respin
This version is both less integrated with systemd and better integrated
with systemd, we work with it, but it's not a hard dependency as it was
previously.  When a parent device is added with mdev support (or the
not-yet-upstream MDEV_STATE uevents are triggered), we can't add a
SYSTEMD_WANTS to the device as we had previously.  Mediated device
support can be added and removed on the parent device, but systemd only
acts on the Wants= when the device first becomes active.  Therefore,
SYSTEMD_WANTS only works once per parent device.  Instead use these
udev events to start persistent mdev devices directly, which in turn
triggers add uevents for the mdev devices, which can then make use of
the Wants= directive as intended to be tracked and potentially managed
via a systemd service.  We also register a systemd tag for mdev devices
and use the systemd device name when starting the device unit instance.
This lets us use BindsTo= so that the service is automatically stopped
if the device is removed.  Systemd interaction is therefore now limited
to the add/remove uevents and use of the service units.

We now support proper argument parsing using getopts.  The first arg
is still the command, but parameters per command are now specified as
args and can be placed in any order.

The commands have been revised.  They've been shortened, since it
seems unnecessary to include 'mdev' in every command for a program
named 'mdevctl'.  The basic command set is now define/undefine/modify,
start/stop, and list.  The first set operate on the config for a
device and imporantly we now support 'transient' devices, which can
be created either externally or via mdevctl.  These devices have no
config, but can be promoted to a defined device using the 'define'
command.  The start/stop commands manage the active set of devices
and can work with defined devices or create transient devices with
a fully specified set of options.  The list command allows us to
see currently running devices, or defined devices, or available mdev
types on the host.

This commit also removes the sbin vs libexec split of the mdevctl
scripts, it's all done in one place now.  The intended user commands
are listed in the usage text with getopts support, the "behind the
scenes commands" are not.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>