Attribute support
Attributes allow post-creation configuration of a device by writing the
specified values to the device attributes.  If a comma separated list
is provided as a value, it's converted to an array in the json
representation and each value is applied in sequential order.  mdevctl
also walks attributes in sequential order, therefore the sequence of
attribute writes can be fully specified.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>