Interpret escapes
We interpret comma separated attribute values as an array, but in
case there's a need to write comma separated values to a sysfs
attribute, interpret escapes in the value.  For instance
"--value 3,4" would be stored as ["3","4"] resulting in separate,
sequential writes of "3" then "4" to the attribute.  If we actually
want "3,4" written to the attribute, it can be escaped as
"--value 3\\\\x2c4", which is stored in the json config file as
"3\\x2c4" and with this modification interpreted as "3,4" on write
to the attribute.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>