Dump and import support
The 'list' command now supports a --dumpjson option which will print
the JSON configuration for a device.  The 'list' command also now
includes UUID and PARENT device specification as this new --dumpjson
option is only available when the list is filtered to a single device.
The --dumpjson option supports either the defined devices or active
devices, however the latter may be incomplete relative to any device
attributes.  Listing of types now also supports the PARENT option to
limit the listing to types for a specific parent device.

The 'define' and 'start' commands now support a --jsonfile option to
allow these to work with a provided configuration file.  A PARENT
device must be specified as this is not provided via the JSON file.
This should allow direct creation of and defining of devices with
attributes specified in the JSON file.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>