simplify attribute handling
Instead of automatically parsing a list of values for an attribute
into an array and then writing each value to the attribute
individually, just write the value given to the attribute as-is.

While this makes specifying lists of values a bit more tedious
(although it is probably better to write a config file directly
for complex configurations anyway), it gives us some benefits:
- no weird escaping if you explicitly want to specify a comma-
  separated list of values for an attribute
- you can delete one value for an attribute explicitly (by index),
  no need to delete the whole range and re-add
- slightly simpler parsing code

Note: existing configs with attributes are not converted.

Signed-off-by: Cornelia Huck <cohuck@redhat.com>