Split mdev types from list command
Per the below issue, it's undesirable that we use -t|--type for various
commands and -t|--types for the list command to get a listing of vendor
mdev types.  A user might conclude that 'list -t=TYPE' should list only
that mdev type rather than the types known to the system.  To resolve
this, let's create a 'types' command and additionally enhance it to
support a JSON output format.  Also, hide the attribute listing for
defined devices behind a verbose option.

Fixes: https://github.com/mdevctl/mdevctl/issues/9
Signed-off-by: Alex Williamson <alex.williamson@redhat.com>