Fix systemd UUID escaping
The escaping of the hyphens in the UUID of the device path is getting
lost between SYSTEMD_WANTS= and the Wants= of the resulting device as
seen in 'systemctl show'.  This means that our unit templat is getting
called with 'x2d' in place of the hyphens rather than '\x2d' and the
systemd-escape unmangling falls apart resulting in failed services.
This is similar to https://github.com/systemd/systemd/issues/7109,
which references a commit which became:

dcebc9bae4dc ("core: when a unit template is specified in SYSTEMD_WANTS=, instantiate it with sysfs path")

This is included since v236 and allows us to simply let systemd fill
in the escaped device path if we only supply the service template.
We can then also remove the mdevctl script code previously necessary
to generate that path.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>