Fix device udev tagging
When an mdev device is created, we get an add uevent, likely closely
followed by a bind uevent.  The add event triggers the systemd service,
but the bind event seems to remove the systemd tag from the device.
Without that systemd tag, a systemd daemon-reload will stop the device
unit, which (thanks to the BindsTo) stops the service unit, which
removes the device.  This is not desirable.  If we change the action
to all but remove, we retain the tag and the device persists across
a daemon-reload.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>