Remove systemd support
Now that it's fixed, remove it.  It's rather hard to justify the
systemd integration, we tag mdevs in order to have systemd create
a dynamic device unit, which only exists so that our template
service unit can bind to it, which really only allows the mdev
to be stopped via a convoluted unit file name involving the
sysfs device path and systemd escaped UUID, which is difficult to
use correctly.  Remove it and see if anyone cares.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>