Makefile: Format the changelog date the legacy way
Since Git >=2.6 it is possible to use a custom date format when
formatting the git history. We need a custom format because of the
overly restrictive format imposed by the RPM %changelog directive.
For platforms like CentOS-7 with Git <2.0 the rpm build naturally
fails. Fox it by introducing an intermediary step invoking the 'date'
command handling the custom format and then simply use the result in
git log formatting.

Signed-off-by: Erik Skultety <eskultet@redhat.com>