Add callout function to mdevctl
Add callout function to validate the device setups and handle device-type
specific operations. The callout function will invoke scripts at various
points during configuration/starting of devices. The callouts will
call out to scripts that are registered for a specific device type and
make sure that the device and the system are sane. There are two types
of callouts: pre-command callout and post-command callout. A notifier is
introduced to invoke notification scripts that can signal other programs
of mdevctl's retults.

The pre-command callout is invoked before any changes is done to the system.
The callouts will call out to scripts to check a device is sane before
starting/configuring to the system. Errors reported during the pre-command
callout are disruptive. The program will exit early if error detected
and no change will be performed.

The post-command callout is invoked after the primary command execution
if the pre-command callout was successful. The callouts will call out to
scripts to check the whole system is sane after a device is started/configured,
or perform cleanups if the primary command execution fails. Errors reported
during the post-command callout are non-disruptive. The error will report
upon script error, and all scripts will be ran.

The notify callout is invoked once at the tail-end of mdevctl to signal
other programs of the mdevctl command's results. Errors reported during
the notify callout are non-disruptive.

Signed-off-by: Zhuoying Cai <zycai@linux.vnet.ibm.com>