Split command code into two case blocks
In prepapartion for the pre and post callout code additions, let's split
the command code into two separate case blocks. The first block is
used for parameter parsing and "prep work" (such as reading / creating
the config file in memory). The second block is used for the "real work"
(such as writing the config to the system, starting or stopping the
device, etc.)

Some commands (such as "list" and "types") are not anticipated to
support callouts in the immediate future, but the groundwork for any
commands that desire to support the callout feature is now in place.

Signed-off-by: Collin Walling <walling@linux.ibm.com>