Add beginning of mdevctl rust conversion
This is the beginning of an experimental rewrite of mdevctl into rust.
The motivation was because of the slow performance of mdevctl,
particularly in handling json data. This is due to the fact that mdevctl
is currently written in shell script and calls out to the command-line
utility `jq` to manipulate json data, often multiple times per command.
The result of this is that when 20 mdevs are defined on my laptop, the
command 'mdevctl list --defined --dumpjson' can take about 7 seconds to
complete.

As of now, the command-line parsing mimics the original mdevctl utility,
but the commands are unimplemented.

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>