tests: run tests with a temp root dir
When testing mdevctl, provide an environment object that is configured
to use a temporary directory to read and write configuration files. This
way the tests will not interact with the system paths on the computer
and will not require special permissions.  This scratch directory is
automatically removed at the end of the test.

New type TestEnvironment encapsulates both the filesystem root and the
test data directory.

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>