refactor define & start
The define and start code blocks may accept a single --jsonfile argument
en lieu of multiple parameters. Since both paths end with the same (or
very close to same) lines of code, let's place the multi-parameter path
in an else block after the --jsonfile block so that both paths can share
the same few lines of code at the end of the command block.

Signed-off-by: Collin Walling <walling@linux.ibm.com>