Change MdevInfo::defined into a function
When a MdevInfo struct doesn't specify a parent, it is not possible to
determine whether it is defined or not. But when a parent is set, the
device may now be defined. So rather than cache this value in a struct
field, convert it to a function so that we don't have to keep it updated
when the 'parent' value changes, etc. This makes the code less
efficient, since it accesses the filesystem everytime we call
is_defined(), but it also makes it more robust.

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>