Add callout support for manual and auto start
This patch only supports manual and auto start.

Move the start_mdev to the second command case-block, and ensure the
return code of start_mdev is captured at the end of mdevclt. The return
code of the start_mdev is important because there might be external scripts
that invoke mdevctl itself and listen success/failure return code when
starting devices.

Invoke callout scripts for auto devices at boot time.

Signed-off-by: Zhuoying Cai <zycai@linux.vnet.ibm.com>