MDev: Use Option for parent, mdev_type
These values might be missing until a device is defined or started, so
use the Option type to indicate that they're missing rather than relying
on empty strings. Also add a couple convenience functions to get the
Option value and convert it to a Result so that errors can be easily
propagated to the caller.

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>