Generate a makefile for installing
Cargo is extremely useful for building and installing rust applications
and their dependencies, but it is a fairly limited build system. It only
has support for installing the single target of the compilation: in this
case, the 'mdevctl' binary. In order to install all of the other support
files, we need a different solution on top of cargo. Using a custom
build script (build.rs), we can generate a Makefile with the appropriate
paths to the build/generated files that we can use for isntallation.

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>