Ensure `list` displays devices in a stable order
The list tests were failing sometimes due to differences in environment.
This is because the child devices were simply inserted into a vector in
the order that they were loaded form the filesystem. Explicitly sort
them by uuid so that the order is deterministic and the output is
testable across platforms.

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>