Add a github workflow for releases
Add two new github workflow jobs that run when a new release is made.

In order to make packaging easier for downstream linux distributions,
we want to include a separate 'vendor' tarball alongside the mdevctl
source. In order to publish this tarball, add a new github workflow that
runs `cargo vendor` and uploads the vendor tarball to the appropriate
release page.

A second job uploads the new release to crates.io using a token that has
been added to the repository secrets.

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>