Don't emit warning for files in /etc/mdevctl.d/scripts.d
mdevctl expects persistent device definitions to be stored in a file
that is located at /etc/mdevctl.d/$PARENT/$UUID. The callout scripts are
stored in /etc/mdevctl.d/scripts.d/. So mdevctl interprets the
'scripts.d' directory as the name of a parent device. Add a special case
to skip this directory when looking up device definitions.

Fixes #52

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>