tests: read stdin in callout test scripts
Callout scripts are intended to be passed a JSON device configuration
string on stdin. For our simple callout tests, we used single-line
test scripts that unconditionally returned a constant response code
(either error or success). This sometimes causes the tests to fail due
to the following error:

    [2021-12-01T20:33:25Z DEBUG mdevctl::callouts] failed to execute callout script "/tmp/mdevctl-testZg8CPd/etc/mdevctl.d/scripts.d/callouts/rc1.sh": Failed to write to stdin of command

        Caused by:
            Broken pipe (os error 32)

What seems to be happening is that mdevctl spawns the callout script and
then attempts to write the JSON device configuration to its stdin pipe.
However, the test scripts are so short that they may have exited before
mdevctl can finish writing to stdin, which results in the command
failing with a broken pipe error. In order to avoid this, make sure that
the test scripts read from stdin before exiting.

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>