start: provide useful hint for parent with wrong case
Since the identifier of a parent device can use letters (e.g.
hexadecimal numbers for PCI), a user might specify the parent with
uppercase letters whn the proper sysfs name is all lowercase. When this
happens, it produces a slightly confusing error message, e.g.:

  Error: No available instances of nvidia-46 on 0000:2B:00.0

To make this a bit more user-friendly, look in the sysfs directory for a
directory that matches the user-provided parent string but only differs
in case, and offer that as a hint, e.g.:

  Error: Parent device '0000:2B:00.0' not found. Did you mean '0000:2b:00.0'?

Also add a test that ensures that the error message for this situation
is different than the situation where the parent device does not exist
at all.

Fixes https://github.com/mdevctl/mdevctl/issues/60

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>