Fix intermittent callout failure during test
There is a race that causes callout scripts to occasionally fail to
execute during the test suite. This is caused by the fact that the test
suite is run in multiple threads and the fact that callout scripts are
copied into a temporary test environment directory. See
https://github.com/golang/go/issues/22315 for a thorough description of
the root cause of a similar problem encountered by the 'go' compiler.

Fixes: #64.

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>