callouts: make sure scripts are sorted
Since we always choose the first matching script for a given device,
they need to be attempted in a predictable order. read_dir() does not
ensure any ordering, so we need to do the sorting ourselves. This way,
an administrator can ensure that a particular script has a higher
priority by naming it e.g. '00-scriptname' vs '99-scriptname'.

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>