List attributes of active mdevs
Use the attributes mdev depending on the mdevs state.
Using --defined/-d will show the defined attributes from the stored
definition and otherwise for active mdevs the attributes are loaded with
the help of the callout scripts as it is done on the define subcommand
with an active mdevs uuid specified.

Signed-off-by: Boris Fiuczynski <fiuczy@linux.ibm.com>