tests: fix test_modify() when auto and manual are both set
This test was supposed to fail when both 'auto' and 'manual' are set to
true. It turns out that this situation is prevented by the CLI parsing,
but the modify_command() function itself was not validating that they
were not both set.

The test failed as expected, but it turns out that it was only failing
because the test was not set up correctly. After setting the test up
correctly, the modify_command() function as actually succeeding, even
though it should not have. Add some logic to return an error if both of
these arguments are set to true.

Fixes: https://github.com/mdevctl/mdevctl/issues/74

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>