callouts: add versioning on callout scripts
Introducing CalloutVersion, CalloutVersions and CalloutScripts to
support versioning of the callout API. Scripts provide via a new
get-capabilities callout on stdout a json which describes the actions
and events it supports. The script receives on stdin a json describing
the actions and events which mdevctl has implemented.
Keep the legacy execution pattern in place.

Signed-off-by: Boris Fiuczynski <fiuczy@linux.ibm.com>