callouts: Add a Callout argument to the invoke callback function
Sometimes the callout functionality is not as simple as the general
case (pre -> main-function -> post -> notify). For example, the modify
operation may need to get attributes via a callout script as part of its
operation. In order to support this, add a 'callout' argument to the
main operation callback function. This allows us to use Callout API
within the main operation.