callouts: Move stdin generation from Callout::invoke_script()
Callout::invoke_script() should not have any knowledge about the message
protocol, so this is not the right spot to generate the json device
description to write to stdin on the callout. Move the responsibility
for setting stdin up a level and add it as an argument to the function.

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>
Reviewed-by: Boris Fiuczynski <fiuczy@linux.ibm.com>