mdev: provide better error message for invalid attr index
When a user tries to add or delete an attribute and provides an invalid
attribute index, we were just returning an error and telling them that
it was an invalid index without indicating anything about what indices
were valid. Provide an initial hint in the error message in these cases
by listing the attributes and their indices. Factor out a method for
printing attributes from the MDev::to_text() function.

Reviewed-by: Boris Fiuczynski <fiuczy@linux.ibm.com>
Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>