callouts: add a check_result_fn to invoke_first_matching_script
Add a function parameter to the invoke_first_matching_script methods
parameters to allow a check of the callout scripts output when
responding correctly. If the check_result_fn function returns OK
invoke_first_matching_script returns with the path and output of the
callout script and if check_result_fn returns an error the searching in
invoke_first_matching_script continues.

Signed-off-by: Boris Fiuczynski <fiuczy@linux.ibm.com>