main: assemble active devices without overwriting
Active devices are created from loading the sysfs information. Autostart
is the only attribute that must be retrieved from the persisted
configuration if it exists. Instead of loading the persisted
configuration into the retrieved active mdev data which overwrite valid
active mdev settings be more surgical and use a new mdev for loading
the configuration and the only copy the autostart attribute to the
active mdev.

Signed-off-by: Boris Fiuczynski <fiuczy@linux.ibm.com>