Move some utility functions to Environment
The main.rs file is large and a bit unfocused. And logically, these
functions fit with the Environment object. For example, we're getting
the list of defined devices for that environment, so making that a
method of Environment makes sense. My vision is that eventually, the
only code left in main will be the parsing of the commandline arguments,
calling other functions, and then formatting output. This moves toward a
future where the core functionality can be exposed in a library as well.

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>