Remove list command wrapper
Rather than having list_command() as a wrapper over another function so
that we can capture the output for tests, just make the list_command()
function take a std::io::Write trait and write to that in the function.
Then the tests can just pass in a buffer and the app itself can pass in
stdout.

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>