mdev: add check in load_from_sysfs for parallel device removal
Since we already checked in the sysfs that mdev exists the failures to
retrieve parent and type can either be caused by
 a) the mdev was delete in the meantime or
 b) an unknown reason
From case a) we can actually recover into a nearly data consistent way
by rechecking if the mdev was deleted. For case b) we must report the
error as before.

Signed-off-by: Boris Fiuczynski <fiuczy@linux.ibm.com>